+++ app/questionnaire/__init__.py
@@ -1,3 +1,4 @@
+from .location import Location
 from .questionnaire_schema import QuestionnaireSchema
 
-__all__ = ["QuestionnaireSchema"]
+__all__ = ["QuestionnaireSchema", "Location"]
